FieldAPI: Move utility classes and methods to dedicated sub-package
