+++ loki/transformations/__init__.py
@@ -18,6 +18,7 @@ from loki.transformations.argument_shape import * # noqa
 from loki.transformations.data_offload import * # noqa
 from loki.transformations.drhook import * # noqa
 from loki.transformations.extract import * # noqa
+from loki.transformations.field_api import * # noqa
 from loki.transformations.hoist_variables import * # noqa
 from loki.transformations.idempotence import * # noqa
 from loki.transformations.inline import * # noqa
