+++ loki/transformations/parallel/field_api.py
@@ -9,18 +9,16 @@
 Transformation utilities to manage and inject FIELD-API boilerplate code.
 """
 
-from enum import Enum
 from loki.expression import symbols as sym
 from loki.ir import (
     nodes as ir, FindNodes, FindVariables, Transformer
 )
-from loki.scope import Scope
 from loki.logging import warning
 from loki.tools import as_tuple
 
+
 __all__ = [
-    'remove_field_api_view_updates', 'add_field_api_view_updates', 'get_field_type',
-    'field_get_device_data', 'field_sync_host', 'FieldAPITransferType'
+    'remove_field_api_view_updates', 'add_field_api_view_updates'
 ]
 
 
@@ -152,85 +150,3 @@ def add_field_api_view_updates(routine, dimension, field_group_types, dim_object
             return loop
 
     routine.body = InsertFieldAPIViewsTransformer().visit(routine.body, scope=routine)
-
-
-def get_field_type(a: sym.Array) -> sym.DerivedType:
-    """
-    Returns the corresponding FIELD API type for an array.
-
-    This function is IFS specific and assumes that the
-    type is an array declared with one of the IFS type specifiers, e.g. KIND=JPRB
-    """
-    type_map = ["jprb",
-                "jpit",
-                "jpis",
-                "jpim",
-                "jpib",
-                "jpia",
-                "jprt",
-                "jprs",
-                "jprm",
-                "jprd",
-                "jplm"]
-    type_name = a.type.kind.name
-
-    assert type_name.lower() in type_map, ('Error array type kind is: '
-                                           f'"{type_name}" which is not a valid IFS type specifier')
-    rank = len(a.shape)
-    field_type = sym.DerivedType(name="field_" + str(rank) + type_name[2:4].lower())
-    return field_type
-
-
-
-class FieldAPITransferType(Enum):
-    READ_ONLY = 1
-    READ_WRITE = 2
-    WRITE_ONLY = 3
-
-
-def field_get_device_data(field_ptr, dev_ptr, transfer_type: FieldAPITransferType, scope: Scope):
-    """
-    Utility function to generate a :any:`CallStatement` corresponding to a Field API
-    ``GET_DEVICE_DATA`` call.
-    
-    Parameters
-    ----------
-    field_ptr: pointer to field object
-        Pointer to the field to call ``GET_DEVICE_DATA`` from.
-    dev_ptr: :any:`Array`
-        Device pointer array
-    transfer_type: :any:`FieldAPITransferType`
-        Field API transfer type to determine which ``GET_DEVICE_DATA`` method to call.
-    scope: :any:`Scope`
-        Scope of the created :any:`CallStatement`
-    """
-    if not isinstance(transfer_type, FieldAPITransferType):
-        raise TypeError(f"transfer_type must be of type FieldAPITransferType, but is of type {type(transfer_type)}")
-    if transfer_type == FieldAPITransferType.READ_ONLY:
-        suffix = 'RDONLY'
-    elif transfer_type == FieldAPITransferType.READ_WRITE:
-        suffix = 'RDWR'
-    elif transfer_type == FieldAPITransferType.WRITE_ONLY:
-        suffix = 'WRONLY'
-    else:
-        suffix = ''
-    procedure_name = 'GET_DEVICE_DATA_' + suffix
-    return ir.CallStatement(name=sym.ProcedureSymbol(procedure_name, parent=field_ptr, scope=scope),
-                            arguments=(dev_ptr.clone(dimensions=None),), )
-
-
-def field_sync_host(field_ptr, scope):
-    """
-    Utility function to generate a :any:`CallStatement` corresponding to a Field API
-    ``SYNC_HOST`` call.
-    
-    Parameters
-    ----------
-    field_ptr: pointer to field object
-        Pointer to the field to call ``SYNC_HOST`` from.
-    scope: :any:`Scope`
-        Scope of the created :any:`CallStatement`
-    """
-
-    procedure_name = 'SYNC_HOST_RDWR'
-    return ir.CallStatement(name=sym.ProcedureSymbol(procedure_name, parent=field_ptr, scope=scope), arguments=())
