+++ loki/transformations/parallel/tests/test_field_api.py
@@ -12,12 +12,15 @@ from loki.frontend import available_frontends, OMNI
 from loki.ir import nodes as ir, FindNodes
 from loki.expression import symbols as sym
 from loki.scope import Scope
-from loki.transformations.parallel import (
-    remove_field_api_view_updates, add_field_api_view_updates,
-    get_field_type, field_get_device_data, FieldAPITransferType
-)
 from loki.types import BasicType, SymbolAttributes
 from loki.logging import WARNING
+
+from loki.transformations.field_api import (
+    get_field_type, field_get_device_data, FieldAPITransferType
+)
+from loki.transformations.parallel import (
+    remove_field_api_view_updates, add_field_api_view_updates
+)
 
 
 @pytest.mark.parametrize('frontend', available_frontends(
